requirements.txt
qrcode[pil]==7.4.2
pillow==10.4.0
python-dotenv==1.0.1
